Issue #4785: DetailNodeTreeStringPrinter#createFakeBlockComment move to CommonUtils
