+++ src/main/java/com/puppycrawl/tools/checkstyle/DetailNodeTreeStringPrinter.java
@@ -29,7 +29,7 @@ import com.puppycrawl.tools.checkstyle.api.DetailNode;
 import com.puppycrawl.tools.checkstyle.api.FileText;
 import com.puppycrawl.tools.checkstyle.api.JavadocTokenTypes;
 import com.puppycrawl.tools.checkstyle.api.LocalizedMessage;
-import com.puppycrawl.tools.checkstyle.api.TokenTypes;
+import com.puppycrawl.tools.checkstyle.utils.CommonUtils;
 import com.puppycrawl.tools.checkstyle.utils.JavadocUtils;
 
 /**
@@ -38,19 +38,21 @@ import com.puppycrawl.tools.checkstyle.utils.JavadocUtils;
  */
 public final class DetailNodeTreeStringPrinter {
 
-    /** OS specific line separator. */
+    /**
+     * OS specific line separator.
+     */
     private static final String LINE_SEPARATOR = System.getProperty("line.separator");
 
-    /** Symbols with which javadoc starts. */
-    private static final String JAVADOC_START = "/**";
-
-    /** Prevent instances. */
+    /**
+     * Prevent instances.
+     */
     private DetailNodeTreeStringPrinter() {
         // no code
     }
 
     /**
      * Parse a file and print the parse tree.
+     *
      * @param file the file to print.
      * @return parse tree as a string
      * @throws IOException if the file could not be read.
@@ -61,6 +63,7 @@ public final class DetailNodeTreeStringPrinter {
 
     /**
      * Parse block comment DetailAST as Javadoc DetailNode tree.
+     *
      * @param blockComment DetailAST
      * @return DetailNode tree
      */
@@ -75,16 +78,18 @@ public final class DetailNodeTreeStringPrinter {
 
     /**
      * Parse javadoc comment to DetailNode tree.
+     *
      * @param javadocComment javadoc comment content
      * @return tree
      */
     private static DetailNode parseJavadocAsDetailNode(String javadocComment) {
-        final DetailAST blockComment = createFakeBlockComment(javadocComment);
+        final DetailAST blockComment = CommonUtils.createFakeBlockComment(javadocComment);
         return parseJavadocAsDetailNode(blockComment);
     }
 
     /**
      * Builds error message base on ParseErrorMessage's message key, its arguments, etc.
+     *
      * @param parseErrorMessage ParseErrorMessage
      * @return error message
      */
@@ -102,9 +107,10 @@ public final class DetailNodeTreeStringPrinter {
 
     /**
      * Print AST.
-     * @param ast the root AST node.
+     *
+     * @param ast        the root AST node.
      * @param rootPrefix prefix for the root node
-     * @param prefix prefix for other nodes
+     * @param prefix     prefix for other nodes
      * @return string AST.
      */
     public static String printTree(DetailNode ast, String rootPrefix, String prefix) {
@@ -130,6 +136,7 @@ public final class DetailNodeTreeStringPrinter {
 
     /**
      * Get indentation for a node.
+     *
      * @param node the DetailNode to get the indentation for.
      * @return the indentation in String format.
      */
@@ -163,43 +170,15 @@ public final class DetailNodeTreeStringPrinter {
 
     /**
      * Parse a file and return the parse tree.
+     *
      * @param file the file to parse.
      * @return the root node of the parse tree.
      * @throws IOException if the file could not be read.
      */
     private static DetailNode parseFile(File file) throws IOException {
         final FileText text = new FileText(file.getAbsoluteFile(),
-            System.getProperty("file.encoding", "UTF-8"));
+                System.getProperty("file.encoding", "UTF-8"));
         return parseJavadocAsDetailNode(text.getFullText().toString());
-    }
-
-    /**
-     * Creates DetailAST block comment to pass it to the Javadoc parser.
-     * @param content comment content.
-     * @return DetailAST block comment
-     */
-    private static DetailAST createFakeBlockComment(String content) {
-        final DetailAST blockCommentBegin = new DetailAST();
-        blockCommentBegin.setType(TokenTypes.BLOCK_COMMENT_BEGIN);
-        blockCommentBegin.setText("/*");
-        blockCommentBegin.setLineNo(0);
-        blockCommentBegin.setColumnNo(-JAVADOC_START.length());
-
-        final DetailAST commentContent = new DetailAST();
-        commentContent.setType(TokenTypes.COMMENT_CONTENT);
-        commentContent.setText("*" + content);
-        commentContent.setLineNo(0);
-        // javadoc should starts at 0 column, so COMMENT_CONTENT node
-        // that contains javadoc identificator has -1 column
-        commentContent.setColumnNo(-1);
-
-        final DetailAST blockCommentEnd = new DetailAST();
-        blockCommentEnd.setType(TokenTypes.BLOCK_COMMENT_END);
-        blockCommentEnd.setText("*/");
-
-        blockCommentBegin.setFirstChild(commentContent);
-        commentContent.setNextSibling(blockCommentEnd);
-        return blockCommentBegin;
     }
 
 }
