+++ src/test/java/com/puppycrawl/tools/checkstyle/DetailNodeTreeStringPrinterTest.java
@@ -34,9 +34,7 @@ import org.junit.Assert;
 import org.junit.Test;
 import org.powermock.reflect.Whitebox;
 
-import com.puppycrawl.tools.checkstyle.api.DetailAST;
 import com.puppycrawl.tools.checkstyle.api.LocalizedMessage;
-import com.puppycrawl.tools.checkstyle.api.TokenTypes;
 
 public class DetailNodeTreeStringPrinterTest extends AbstractTreeTestSupport {
 
@@ -76,31 +74,6 @@ public class DetailNodeTreeStringPrinterTest extends AbstractTreeTestSupport {
             assertEquals("Generated and expected parse error messages don't match",
                     expected, ex.getMessage());
         }
-    }
-
-    @Test
-    public void testCreationOfFakeCommentBlock() throws Exception {
-        final Method createFakeBlockComment =
-                Whitebox.getMethod(DetailNodeTreeStringPrinter.class,
-                        "createFakeBlockComment", String.class);
-
-        final DetailAST testCommentBlock =
-                (DetailAST) createFakeBlockComment.invoke(null, "test_comment");
-        assertEquals("Invalid token type",
-                TokenTypes.BLOCK_COMMENT_BEGIN, testCommentBlock.getType());
-        assertEquals("Invalid text", "/*", testCommentBlock.getText());
-        assertEquals("Invalid line number", 0, testCommentBlock.getLineNo());
-
-        final DetailAST contentCommentBlock = testCommentBlock.getFirstChild();
-        assertEquals("Invalid tiken type",
-                TokenTypes.COMMENT_CONTENT, contentCommentBlock.getType());
-        assertEquals("Invalid text", "*test_comment", contentCommentBlock.getText());
-        assertEquals("Invalid line number", 0, contentCommentBlock.getLineNo());
-        assertEquals("Invalid column number", -1, contentCommentBlock.getColumnNo());
-
-        final DetailAST endCommentBlock = contentCommentBlock.getNextSibling();
-        assertEquals("Invalid tiken type", TokenTypes.BLOCK_COMMENT_END, endCommentBlock.getType());
-        assertEquals("Invalid text", "*/", endCommentBlock.getText());
     }
 
     @Test
